style: rename constants.ts -> .tsx
